// web/src/api/generated/model/issueSummary.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Condux Control Plane API
 * OpenAPI spec version: v1
 */

export interface IssueSummary {
  id: string;
  projectId: number;
  fingerprint: string;
  title: string;
  culprit: string;
  level: number;
  status: number;
  firstSeen: string;
  lastSeen: string;
  eventCount: number;
  /** @nullable */
  assigneeUserId: number | null;
  /** @nullable */
  firstRelease: string | null;
}